Pre  officia release branch sync with master branch

+++ Ginger/GingerCoreCommon/VariablesLib/DecimalRandom.cs
@@ -75,7 +75,7 @@ namespace GingerCore.Variables
                 d0_1 = d0_1 / 10M;
             }
 
-            decimal d = d0_1 * (decimal)Math.Abs(Max - Min) + Min;
+            decimal d = d0_1 * Math.Abs(Max - Min) + Min;
             if (integer)
             {
                 d = Math.Round(d, 0);
